pyproject.toml
[project]
name = "helath-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-ninja>=1.3.0",
    "logfire[asgi,django]>=2.11.0",
    "model-bakery>=1.20.1",
    "psycopg-binary>=3.2.3",
    "psycopg[pool]>=3.2.3",
    "pytest-asyncio>=0.25.1",
    "pytest-django>=4.9.0",
    "pytest-faker>=2.0.0",
    "uvicorn>=0.34.0",
]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.main.settings"
python_files = "tests/* test_*.py *_tests.py"
